Add option to show error and fix cut off for docs

+++ src/qa/indexer/ValueComparator.java
@@ -5,9 +5,9 @@ import java.util.Map;
 
 public class ValueComparator implements Comparator<String> {
 
-	Map<String, Integer> base;
+	Map<String, Float> base;
 
-	public ValueComparator(Map<String, Integer> base) {
+	public ValueComparator(Map<String, Float> base) {
 		this.base = base;
 	}
